+++ vs/Vector.h
@@ -17,7 +17,7 @@ public:
 	void normalise();
 
 	Vector& operator=(const Vector& rhs);
-	Vector operator*(float scalar) const;
+	Vector operator*(double scalar) const;
 	Vector operator-() const;
 	Vector operator-(const Vector& rhs) const;
 	Vector operator+(const Vector& rhs) const;
